set date picker for reminder
